+++ pkg/assembler/backends/inmem/builder.go
@@ -20,9 +20,8 @@ import (
 	"errors"
 	"strconv"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 type builderMap map[string]*builderStruct
