+++ pkg/assembler/backends/inmem/certifyScorecard.go
@@ -21,9 +21,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 // Internal data: link between source and scorecard (certifyScorecard)
