+++ pkg/assembler/backends/inmem/ghsa.go
@@ -22,9 +22,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 // TODO: convert to unit test
