+++ pkg/assembler/backends/inmem/hashEqual.go
@@ -21,10 +21,9 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/guacsec/guac/pkg/assembler/graphql/model"
 	"github.com/vektah/gqlparser/v2/gqlerror"
 	"golang.org/x/exp/slices"
-
-	"github.com/guacsec/guac/pkg/assembler/graphql/model"
 )
 
 // Internal hashEqual
