+++ pkg/assembler/backends/inmem/isDependency.go
@@ -19,9 +19,8 @@ import (
 	"context"
 	"strconv"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 // Internal data: link between packages and dependent packages (isDependency)
