+++ pkg/assembler/backends/inmem/path.go
@@ -19,9 +19,8 @@ import (
 	"context"
 	"strconv"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 type edgeMap map[model.Edge]bool
