+++ pkg/assembler/backends/inmem/pkg.go
@@ -22,9 +22,8 @@ import (
 	"reflect"
 	"strconv"
 
-	"github.com/vektah/gqlparser/v2/gqlerror"
-
 	"github.com/guacsec/guac/pkg/assembler/graphql/model"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 )
 
 // TODO: move this into a unit test for this file
