+++ pkg/assembler/backends/inmem/pkgEqual.go
@@ -19,10 +19,9 @@ import (
 	"context"
 	"strconv"
 
+	"github.com/guacsec/guac/pkg/assembler/graphql/model"
 	"github.com/vektah/gqlparser/v2/gqlerror"
 	"golang.org/x/exp/slices"
-
-	"github.com/guacsec/guac/pkg/assembler/graphql/model"
 )
 
 type pkgEqualList []*pkgEqualStruct
